cfg-c, cfg-go: Generalized switch-handling code to handle both languages
